# portal/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Портал Социальных аптек API
  version: 1.0.0
  description: |
    Контракт REST API портала. Аутентификация выполняется HTTP-only cookie `X-Access-Token`;
    обновление — cookie `X-Refresh-Token`. Для каждого POST, PUT, PATCH и DELETE передавайте
    cookie `XSRF-TOKEN` и одноимённый заголовок `X-XSRF-TOKEN`.
servers:
  - url: /
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/user-groups:
    get: { tags: [Группы пользователей], summary: Получить группы пользователей, parameters: [ { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/UserGroupList' } } }
    post: { tags: [Группы пользователей], summary: Создать группу пользователей, parameters: [ { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserGroupRequest' } } } }, responses: { '201': { $ref: '#/components/responses/UserGroup' } } }
  /api/v1/user-groups/{id}:
    get: { tags: [Группы пользователей], summary: Получить группу пользователей, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/UserGroup' } } }
    put: { tags: [Группы пользователей], summary: Обновить группу пользователей, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserGroupUpdateRequest' } } } }, responses: { '200': { $ref: '#/components/responses/UserGroup' } } }
    delete: { tags: [Группы пользователей], summary: Удалить группу пользователей, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { name: confirm, in: query, schema: { type: boolean, default: false } }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа удалена } } }
...
  /api/v1/user-groups/safe-delete/{id}:
    patch: { tags: [Группы пользователей], summary: Безопасно удалить группу, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа деактивирована } } }
  /api/v1/user-groups/restore/{id}:
    patch: { tags: [Группы пользователей], summary: Восстановить группу, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа восстановлена } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: X-Access-Token
      description: JWT access token. Cookie `X-Refresh-Token` используется только `/api/v1/auth/refresh`.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Значение должно совпадать с cookie `XSRF-TOKEN`; обязательно для unsafe запросов.
  parameters:
    CsrfHeader: { name: X-XSRF-TOKEN, in: header, required: true, schema: { type: string }, description: Значение cookie XSRF-TOKEN }
...
    IntegerId: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
...
    IsActive: { name: isActive, in: query, schema: { type: boolean, default: true } }
...
  responses:
...
    UserGroup: { description: Группа пользователей, content: { application/json: { schema: { $ref: '#/components/schemas/UserGroup' } } } }
    UserGroupList: { description: Список групп пользователей, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UserGroup' } } } } }
...
  schemas:
...
    UserGroupRequest:
      type: object
      required: [name]
      description: |
        Запрос создания группы пользователей.
        `visibleGroups` — id групп, которым эта группа будет видна.
      properties:
        name: { type: string, minLength: 1, description: Наименование группы }
        phoneNumber: { type: string, pattern: '^\+?[0-9]{7,15}$', description: Телефон группы, например +79001234567 }
        internalNumber: { type: string, pattern: '^[0-9]{1,20}$', description: Внутренний номер }
        extensionNumber: { type: string, pattern: '^[0-9]{1,20}$', description: Добавочный номер }
        visibleGroups:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1 }
          description: Список id видимых групп пользователей
      example:
        name: IT-отдел
        phoneNumber: "+79001234567"
        internalNumber: "100"
        extensionNumber: "10"
        visibleGroups: [1, 2]
    UserGroupUpdateRequest:
      type: object
      description: Запрос изменения группы пользователей (все поля опциональны)
      properties:
        name: { type: string, minLength: 1, description: Наименование группы }
        phoneNumber: { type: string, pattern: '^\+?[0-9]{7,15}$', description: Телефон группы }
        internalNumber: { type: string, pattern: '^[0-9]{1,20}$', description: Внутренний номер }
        extensionNumber: { type: string, pattern: '^[0-9]{1,20}$', description: Добавочный номер }
        visibleGroups:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1 }
          description: Список id видимых групп пользователей
    UserGroup:
      type: object
      description: Группа пользователей
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string }
        phoneNumber: { type: string, nullable: true }
        internalNumber: { type: string, nullable: true }
        extensionNumber: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        isActive: { type: boolean }
...